requirements.txt
### These are the Libraries needed to run the app ###
joblib
pandas
numpy
sklearn
streamlit
Pillow
scipy
flask
pymongo
pymongo[srv]